Renamed some files and applied filter to others
