+++ src/main.cpp
@@ -9,5 +9,5 @@ int main(int argCount, char **argValues) {
     }
 
     os.Terminate();
-	return 0;
+    return 0;
 }
