+++ src/util/util.cpp
@@ -1,4 +1,4 @@
-#include "util/UtilType.hpp"
+#include "util/utiltype.hpp"
 
 namespace trillek {
 namespace util {
